pipeline-manager: fix SQL compiler terminating upon cancel
The compilation process is now put into a process group, such that it's
possible to terminate it with all of its spawned subprocesses. This is
done using the `ProcessGroupTerminator` which terminates the process
group when going out of scope. This is the same approach as used for the
Rust compiler.
